Harden API configuration and error handling
- add typed JWT settings validation and update package versions

- make exception middleware resilient when database logging fails

- align ZonaGarita identifiers to byte to fix invalid cast issues

+++ LiquidacionPeajesNew.Application/DTOs/Requests/GaritaRequest.cs
@@ -4,7 +4,7 @@
     {
         public string NombreGarita { get; set; }
         public string RucProveedor { get; set; }
-        public short IdZonaGarita { get; set; }
+        public byte IdZonaGarita { get; set; }
         public int UbicacionGarita { get; set; }
         public bool IdaVuelta { get; set; }
         public short IdTipoGarita { get; set; }
